add changes

+++ database.sql
@@ -8,7 +8,7 @@ CREATE TABLE IF NOT EXISTS urls (
 CREATE TABLE IF NOT EXISTS urls_checks (
         id bigint PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
         url_id bigint REFERENCES urls (id),
-        status_code smallint,
+        status_code int,
         h1 varchar(255),
         title varchar(255),
         description text,
